+++ php-classes/RoomManager.php
@@ -51,6 +51,12 @@
       if (!is_numeric($id))
         return "Id is not a valid number.";
 
+      $sql = "SELECT id FROM room WHERE id='$id'";
+      $response = $this->pdo->sqlRecord($sql);
+    
+      if (empty($response) || $response == NULL)
+        return "There is no room with that id.";
+
       $sql = "DELETE FROM room WHERE id='$id'";
       $response = $this->pdo->sqlQuery($sql);
 
@@ -81,6 +87,12 @@
       if (!is_string($name))
         return "Name is not a valid text.";
 
+      $sql = "SELECT id FROM room WHERE id='$id'";
+      $response = $this->pdo->sqlRecord($sql);
+    
+      if (empty($response) || $response == NULL)
+        return "There is no room with that id.";
+
       $sql = "SELECT id FROM room WHERE name='$name'";
       $valid_name_response = $this->pdo->sqlTable($sql);
   
@@ -105,6 +117,9 @@
       $sql = "SELECT * FROM room";
       $response = $this->pdo->sqlTable($sql);
 
+      if (empty($response) || $response == NULL)
+        return "There are no rooms.";
+
       return $response;
     }
 
@@ -121,6 +136,9 @@
 
       $sql = "SELECT * FROM room WHERE id='$id'";
       $response = $this->pdo->sqlRecord($sql);
+
+      if (empty($response) || $response == NULL)
+        return "There is no room with that id.";
   
       return $response;
     }
@@ -140,6 +158,9 @@
 
       $sql = "SELECT * FROM room WHERE name='$name'";
       $response = $this->pdo->sqlRecord($sql);
+
+      if (empty($response) || $response == NULL)
+        return "There is no room with that name.";
 
       return $response;
     }
